requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
fuzzywuzzy
python-levenshtein
